requirements.txt
flask==3.1.0
openai==2.32.0
python-dotenv==1.2.2
gunicorn==23.0.0
langchain-neo4j==0.9.0
langchain-openai
langchain-core
